+++ src/main/java/org/hihan/girinoscope/ui/Checksum.java
@@ -1,0 +1,38 @@
+package org.hihan.girinoscope.ui;
+
+import java.io.IOException;
+import java.io.InputStream;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+public class Checksum {
+
+    public static byte[] createChecksum(Path path) throws IOException, NoSuchAlgorithmException {
+        try ( InputStream input = Files.newInputStream(path)) {
+            MessageDigest messageDigest = MessageDigest.getInstance("MD5");
+            byte[] buffer = new byte[1024];
+            int numRead;
+            while ((numRead = input.read(buffer)) != -1) {
+                messageDigest.update(buffer, 0, numRead);
+            }
+            return messageDigest.digest();
+        }
+    }
+
+    // See https://stackoverflow.com/questions/9655181/how-to-convert-a-byte-array-to-a-hex-string-in-java
+    public static String bytesToHex(byte[] bytes) {
+        char[] hexArray = "0123456789ABCDEF".toCharArray();
+        char[] hexChars = new char[bytes.length * 2];
+        for (int j = 0; j < bytes.length; j++) {
+            int v = bytes[j] & 0xFF;
+            hexChars[j * 2] = hexArray[v >>> 4];
+            hexChars[j * 2 + 1] = hexArray[v & 0x0F];
+        }
+        return new String(hexChars);
+    }
+
+    private Checksum() {
+    }
+}